libp2p: Implement a default response
Implement a default response for incoming requests that don't have a
registered listener that could respond to them. This avoids cases
where the received request times out even if the response is a don't
care for the upper application.
Also changes the implementation of the `Message` trait to use the
`MessageTypeId` `enum` as the type for the `Message` `TYPE_ID` instead
of using a raw `u64`. This also centralizes all different IDs such
that no collission could be made.

+++ network-libp2p/src/dispatch/codecs/typed.rs
@@ -332,7 +332,7 @@ impl<M: Message> Encoder<&M> for MessageCodec {
     type Error = Error;
 
     fn encode(&mut self, message: &M, dst: &mut BytesMut) -> Result<(), Error> {
-        let mut header = Header::new(M::TYPE_ID);
+        let mut header = Header::new(M::TYPE_ID as u64);
         let message_length = Header::SIZE + message.serialized_size();
         header.length = message_length as u32;
